# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kitelepítési Támogató Rendszer API",
        "description": "QR-kódos kitelepítési, regisztrációs és befogadóhelyi nyomon követő prototípus REST API-ja.\n\nAz autentikáció Laravel Sanctum SPA session-cookie alapú: előbb GET /sanctum/csrf-cookie, utána POST /api/login. A böngészőn kívüli kliensek (pl. ez a Swagger UI) ezért a `/api/login` végpontot csak akkor tudják sikeresen meghívni, ha a kliens ugyanabból az origin-ből fut, ahonnan a cookie-kat is fogadja.",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Helyi fejlesztői szerver"
        }
    ],
    "paths": {
...
        "/api/events/{event}/persons/municipality-summary": {
            "get": {
                "tags": [
                    "Persons"
                ],
                "summary": "Regisztráltak száma lakóhely (település) szerint aggregálva, térképes megjelenítéshez",
                "description": "Csak a koordinátával rendelkező településeket adja vissza. Interreg tanulmány \"Geografikus Nyomon Követési Dashboard\" funkciója.",
                "operationId": "c39af08df3f782bb4bd1ef69749e2f68",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Településenkénti összesítés koordinátákkal"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            }
        },
        "/api/events/{event}/persons/bulk-import": {
            "post": {
                "tags": [
                    "Persons"
                ],
                "summary": "Lakossági/önkormányzati névlista tömeges importálása előzetes QR-kód generálással",
                "description": "Interreg tanulmány \"Előzetes kódgenerálás és kiosztás\" funkciója: egy előre összeállított lakossági vagy önkormányzati CSV-lista alapján tömegesen létrehozza a regisztrációkat és azonnal QR-kódot is generál mindegyikhez, hogy a tényleges kitelepítéskor már csak a kód beolvasása legyen szükséges. CSV-fejléc: vezetéknév,keresztnév,okmányszám,település,telefon. Admin, vezető vagy regisztrátor szerepkör szükséges, csak aktív eseményhez.",
                "operationId": "70e4a11f0b59c361874817ab377e6cbb",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Import összegzés: létrehozott személyek QR-kódokkal, hibás sorok"
                    },
                    "403": {
                        "description": "Nincs jogosultság"
                    },
                    "422": {
                        "description": "Validációs hiba vagy nem aktív esemény"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            }
        },
        "/api/persons/{person}": {
            "get": {
                "tags": [
                    "Persons"
                ],
                "summary": "Személy adatlap (regisztráció, speciális igények, állatok, QR-tokenek)",
                "operationId": "a82703e911372267ac34d57b2a6066ad",
                "parameters": [
                    {
                        "name": "person",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Személy adatai"
                    },
                    "404": {
                        "description": "Nincs ilyen személy"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Persons"
                ],
                "summary": "Személy adatainak módosítása",
                "description": "Csak nem lezárt eseményhez tartozó személy módosítható; admin vagy regisztrátor szerepkör szükséges.",
                "operationId": "d89c5420877e0b35a910e7001985885b",
                "parameters": [
                    {
                        "name": "person",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Frissített személy adatai"
                    },
                    "403": {
                        "description": "Nincs jogosultság vagy lezárt esemény"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            }
        },
...
        "/api/persons/{person}/status-history": {
            "get": {
                "tags": [
                    "Persons"
                ],
                "summary": "A személy regisztrációjának státusztörténete időrendben",
                "operationId": "e4ce5d432903c9fe0a494be584b54093",
                "parameters": [
                    {
                        "name": "person",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Státusztörténet bejegyzések"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctumSession": {
                "type": "apiKey",
                "description": "Laravel Sanctum SPA session-cookie autentikáció. Bejelentkezés: POST /api/login.",
                "name": "laravel_session",
                "in": "cookie"
            }
        }
    },
...
}